Add files via upload
File 1: Simple calculator. Numbers are decided by user input and are not part of the code.
File 2: Checks if entered word is a palindrome or not.
File 3: A program to roll 1-4 dice. No. of dice depends on user input.
File 4: A program which converts units of temperature measurement: Celsius, Fahrenheit and Kelvin based on user input.